pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "arxivbuddy"
version = "0.1.0"
description = "L'IA qui lit les papiers de recherche pour toi (et te les traduit en humain)"
authors = [
    { name = "Lemasson Patrick", email = "l.patrick74210@outlook.com" }
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "crewai==0.114.0",
    "arxiv>=1.4.8",
    "python-dotenv>=1.0.0",
    "argparse>=1.4.0",
    "litellm>=1.30.0",
    "langchain>=0.0.335",
    "sentence-transformers>=2.2.2"
]

[project.urls]
Home = "https://github.com/iapourtous/ArxivBuddy"

[tool.setuptools]
# Indique que les paquets sont dans src/
package-dir = {"" = "src"}
packages = ["arxivbuddy", "lib"]

[project.scripts]
arxivbuddy = "arxivbuddy.cli:main"